商户模型增加字段
商户模型增加字段

+++ src/main/java/com/mmnttech/ma/merchant/server/database/mappers/MerchantMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface MerchantMapper {
-    long countByExample(MerchantExample example);
+    int countByExample(MerchantExample example);
 
     int deleteByExample(MerchantExample example);
